fix: do App component as the global gf-26
