+++ apps/frontend/src/index.tsx
@@ -12,6 +12,7 @@ import {
 import { AppRoute } from "~/libs/enums/enums.js";
 import { store } from "~/libs/modules/store/store.js";
 import { Auth } from "~/pages/auth/auth.jsx";
+import { DefaultPage } from "~/pages/default/default.jsx";
 import { NotFound } from "~/pages/not-found/not-found.jsx";
 
 createRoot(document.querySelector("#root") as HTMLElement).render(
@@ -20,20 +21,26 @@ createRoot(document.querySelector("#root") as HTMLElement).render(
 			<RouterProvider
 				routes={[
 					{
-						element: (
-							<ProtectedRoute>
-								<App />
-							</ProtectedRoute>
-						),
+						children: [
+							{
+								element: (
+									<ProtectedRoute>
+										<DefaultPage />
+									</ProtectedRoute>
+								),
+								path: AppRoute.ROOT,
+							},
+							{
+								element: <Auth />,
+								path: AppRoute.SIGN_IN,
+							},
+							{
+								element: <Auth />,
+								path: AppRoute.SIGN_UP,
+							},
+						],
+						element: <App />,
 						path: AppRoute.ROOT,
-					},
-					{
-						element: <Auth />,
-						path: AppRoute.SIGN_IN,
-					},
-					{
-						element: <Auth />,
-						path: AppRoute.SIGN_UP,
 					},
 					{
 						element: <NotFound />,
@@ -41,6 +48,7 @@ createRoot(document.querySelector("#root") as HTMLElement).render(
 					},
 				]}
 			/>
+
 			<ToastContainer />
 		</StoreProvider>
 	</StrictMode>,
