+++ apps/frontend/src/libs/components/app/app.tsx
@@ -1,82 +1,22 @@
-import {
-	Header,
-	Link,
-	Loader,
-	RouterOutlet,
-	Table,
-} from "~/libs/components/components.js";
-import {
-	mockColumns as mockTableColumns,
-	mockData as mockTableData,
-	type Person,
-} from "~/libs/components/table/mock-data.js";
-import { AppRoute } from "~/libs/enums/enums.js";
+import { Loader, RouterOutlet } from "~/libs/components/components.js";
 import {
 	useAppDispatch,
 	useAppSelector,
 	useEffect,
-	useLocation,
 } from "~/libs/hooks/hooks.js";
 import { actions as authActions } from "~/modules/auth/auth.js";
-import { actions as userActions } from "~/modules/users/users.js";
 
 const App = (): JSX.Element => {
 	const dispatch = useAppDispatch();
-	const { pathname } = useLocation();
 	const dataStatus = useAppSelector(({ users }) => users.dataStatus);
-	const users = useAppSelector(({ users }) => users.users);
 
-	const isRoot = pathname === AppRoute.ROOT;
 	const isLoading = dataStatus === "pending";
 
 	useEffect(() => {
 		void dispatch(authActions.getAuthenticatedUser());
 	}, [dispatch]);
 
-	useEffect(() => {
-		if (isRoot) {
-			void dispatch(userActions.loadAll());
-		}
-	}, [isRoot, dispatch]);
-
-	return (
-		<>
-			<Header />
-			<ul>
-				<li>
-					<Link to={AppRoute.ROOT}>Root</Link>
-				</li>
-				<li>
-					<Link to={AppRoute.SIGN_IN}>Sign in</Link>
-				</li>
-				<li>
-					<Link to={AppRoute.SIGN_UP}>Sign up</Link>
-				</li>
-			</ul>
-			<p>Current path: {pathname}</p>
-
-			<div>
-				<RouterOutlet />
-			</div>
-			{isRoot && (
-				<>
-					<h2>Users:</h2>
-					<h3>Status: {dataStatus}</h3>
-
-					{isLoading ? (
-						<Loader />
-					) : (
-						<ul>
-							{users.map((user) => (
-								<li key={user.id}>{user.email}</li>
-							))}
-						</ul>
-					)}
-					<Table<Person> columns={mockTableColumns} data={mockTableData} />
-				</>
-			)}
-		</>
-	);
+	return <>{isLoading ? <Loader /> : <RouterOutlet />}</>;
 };
 
 export { App };
