+++ apps/frontend/src/pages/default/default.tsx
@@ -1,0 +1,58 @@
+import { Header, Loader, Table } from "~/libs/components/components.js";
+import {
+	mockColumns as mockTableColumns,
+	mockData as mockTableData,
+	type Person,
+} from "~/libs/components/table/mock-data.js";
+import { AppRoute } from "~/libs/enums/enums.js";
+import {
+	useAppDispatch,
+	useAppSelector,
+	useEffect,
+	useLocation,
+} from "~/libs/hooks/hooks.js";
+import { actions as userActions } from "~/modules/users/users.js";
+
+const DefaultPage = (): JSX.Element => {
+	const dispatch = useAppDispatch();
+	const { pathname } = useLocation();
+	const dataStatus = useAppSelector(({ users }) => users.dataStatus);
+	const users = useAppSelector(({ users }) => users.users);
+
+	const isRoot = pathname === AppRoute.ROOT;
+	const isLoading = dataStatus === "pending";
+
+	useEffect(() => {
+		if (isRoot) {
+			void dispatch(userActions.loadAll());
+		}
+	}, [isRoot, dispatch]);
+
+	return (
+		<>
+			<Header />
+
+			<p>Current path: {pathname}</p>
+
+			{isRoot && (
+				<>
+					<h2>Users:</h2>
+					<h3>Status: {dataStatus}</h3>
+
+					{isLoading ? (
+						<Loader />
+					) : (
+						<ul>
+							{users.map((user) => (
+								<li key={user.id}>{user.email}</li>
+							))}
+						</ul>
+					)}
+					<Table<Person> columns={mockTableColumns} data={mockTableData} />
+				</>
+			)}
+		</>
+	);
+};
+
+export { DefaultPage };
